# Conflicts:
#	TheLottoApp/Controllers/UserManagementController.cs
#	TheLottoApp/Models/NumberGeneratorViewModel.cs
#	TheLottoApp/TheLottoApp.csproj

+++ TheLottoApp/PyFiles/stastic_engine.py
@@ -34,7 +34,7 @@ class get_the_statistics:
             total_repeated = len(k) - len(set(k))
             prev_dict[total_repeated] +=1
         for j in list(prev_dict):
-            prev_dict[j] = round((prev_dict[j]/len(self.results))*100,2)
+            prev_dict[j] = round((prev_dict[j]/float(len(self.results)))*100,2)
         return prev_dict
     def get_nb_of_odds(self):
         odd_dict = {0:0, 1:0, 2:0, 3:0, 4:0, 5:0, 6:0, 7:0}
@@ -42,7 +42,7 @@ class get_the_statistics:
             k = sum([int(x)%2==1 for x in result.numbers.split(',')])
             odd_dict[k] +=1
         for i in list(odd_dict):
-            odd_dict[i] = round((odd_dict[i]/len(self.results))*100,2)
+            odd_dict[i] = round((odd_dict[i]/float(len(self.results)))*100,2)
         return odd_dict          
         
     def get_ranges(self):
@@ -55,7 +55,7 @@ class get_the_statistics:
             m = sum([int(x)>30 for x in result.numbers.split(',')])
             range_dict['Above 30'] += m
         for i in list(range_dict):
-            range_dict[i] = round(range_dict[i]/len(self.results),1) 
+            range_dict[i] = round(range_dict[i]/float(len(self.results)),1) 
         return range_dict
         
     def get_score_chart(self):
@@ -64,8 +64,8 @@ class get_the_statistics:
             tot_score = 0
             for ball in result.numbers.split(','):
                 tot_score = tot_score + self.score_dict[ball]
-            tot_score = tot_score / len(result.numbers.split(','))
-            tot_score = tot_score / (len(self.results) * 7)
+            tot_score = tot_score / float(len(result.numbers.split(',')))
+            tot_score = tot_score / float(len(self.results) * 7)
             scores.append(tot_score+0)
         scores.sort()
         return scores
@@ -75,8 +75,8 @@ class get_the_statistics:
             tot_score = 0
             for ball in result.numbers.split(','):
                 tot_score = tot_score + self.score_dict[ball]
-            tot_score = tot_score / len(result.numbers.split(','))
-            tot_score = tot_score / (len(self.results) * 7)
+            tot_score = tot_score / float(len(result.numbers.split(',')))
+            tot_score = tot_score / float((len(self.results) * 7))
             scores.append(tot_score+0)
         return scores
